requirements.txt
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
filelock==3.8.0
Flask==2.2.2
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.1
numpy==1.23.2
packaging==21.3
pyparsing==3.0.9
PyYAML==6.0
redis==4.3.4
regex==2022.8.17
requests==2.28.1
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.0
transformers==4.21.2
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
